NOJIRA move remaining cli tests into correct package
