+++ cli/src/test/kotlin/uk/gov/justice/digital/cli/client/BlockingDomainClientTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.client
+package uk.gov.justice.digital.cli.client
 
 import io.micronaut.context.ApplicationContext
 import io.micronaut.context.annotation.Requires
